+++ test/test__parse.py
@@ -1,5 +1,4 @@
 
-from pe._grammar import Grammar
 from pe.operators import (
     Dot,
     Literal,
@@ -18,97 +17,98 @@ from pe.operators import (
 from pe._parse import loads
 
 
+def eloads(s):
+    start, defmap = loads(s)
+    return defmap[start]
+
+
 def test_loads_dot():
-    assert loads('.') == Dot()
-    assert loads('.  # comment') == Dot()
+    assert eloads('.') == Dot()
+    assert eloads('.  # comment') == Dot()
 
 
 def test_loads_literal():
-    assert loads('"foo"') == Literal('foo')
-    assert loads('"foo"  # comment') == Literal('foo')
-    assert loads('"\\t"') == Literal('\t')
-    assert loads('"\\n"') == Literal('\n')
-    assert loads('"\\v"') == Literal('\v')
-    assert loads('"\\f"') == Literal('\f')
-    assert loads('"\\r"') == Literal('\r')
-    assert loads('"\\""') == Literal('"')
-    assert loads("'\\''") == Literal("'")
-    assert loads("'\\-'") == Literal("-")
-    assert loads("'\\['") == Literal("[")
-    assert loads("'\\\\'") == Literal("\\")
-    assert loads("'\\]'") == Literal("]")
+    assert eloads('"foo"') == Literal('foo')
+    assert eloads('"foo"  # comment') == Literal('foo')
+    assert eloads('"\\t"') == Literal('\t')
+    assert eloads('"\\n"') == Literal('\n')
+    assert eloads('"\\v"') == Literal('\v')
+    assert eloads('"\\f"') == Literal('\f')
+    assert eloads('"\\r"') == Literal('\r')
+    assert eloads('"\\""') == Literal('"')
+    assert eloads("'\\''") == Literal("'")
+    assert eloads("'\\-'") == Literal("-")
+    assert eloads("'\\['") == Literal("[")
+    assert eloads("'\\\\'") == Literal("\\")
+    assert eloads("'\\]'") == Literal("]")
     # TODO: octal, utf8, utf16, utf32, escape errors
 
 
 def test_loads_class():
-    assert loads('[xyz]') == Class('xyz')
-    assert loads('[xyz]  # comment') == Class('xyz')
+    assert eloads('[xyz]') == Class('xyz')
+    assert eloads('[xyz]  # comment') == Class('xyz')
 
 
 def test_loads_nonterminal():
-    assert loads('foo') == Nonterminal('foo')
-    assert loads('foo  # comment') == Nonterminal('foo')
+    assert eloads('foo') == Nonterminal('foo')
+    assert eloads('foo  # comment') == Nonterminal('foo')
 
 
 def test_loads_optional():
-    assert loads('"a"?') == Optional('a')
-    assert loads('"a"?  # comment') == Optional('a')
+    assert eloads('"a"?') == Optional('a')
+    assert eloads('"a"?  # comment') == Optional('a')
 
 
 def test_loads_star():
-    assert loads('"a"*') == Star('a')
-    assert loads('"a"*  # comment') == Star('a')
+    assert eloads('"a"*') == Star('a')
+    assert eloads('"a"*  # comment') == Star('a')
 
 
 def test_loads_plus():
-    assert loads('"a"+') == Plus('a')
-    assert loads('"a"+  # comment') == Plus('a')
+    assert eloads('"a"+') == Plus('a')
+    assert eloads('"a"+  # comment') == Plus('a')
 
 
 def test_loads_sequence():
-    assert loads('"a" "b"') == Sequence('a', 'b')
-    assert loads('"a" "b"  # comment') == Sequence('a', 'b')
+    assert eloads('"a" "b"') == Sequence('a', 'b')
+    assert eloads('"a" "b"  # comment') == Sequence('a', 'b')
 
 
 def test_loads_choice():
-    assert loads('"a" / "b"') == Choice('a', 'b')
-    assert loads('"a" / "b"  # comment') == Choice('a', 'b')
+    assert eloads('"a" / "b"') == Choice('a', 'b')
+    assert eloads('"a" / "b"  # comment') == Choice('a', 'b')
 
 
 def test_loads_and():
-    assert loads('&"a"') == And('a')
-    assert loads('&"a"  # comment') == And('a')
+    assert eloads('&"a"') == And('a')
+    assert eloads('&"a"  # comment') == And('a')
 
 
 def test_loads_not():
-    assert loads('!"a"') == Not('a')
-    assert loads('!"a"  # comment') == Not('a')
+    assert eloads('!"a"') == Not('a')
+    assert eloads('!"a"  # comment') == Not('a')
 
 
 def test_loads_raw():
-    assert loads('~"a"') == Raw('a')
-    assert loads('~"a"  # comment') == Raw('a')
+    assert eloads('~"a"') == Raw('a')
+    assert eloads('~"a"  # comment') == Raw('a')
 
 
 def test_loads_bind():
-    assert loads('x:"a"') == Bind('a', name='x')
-    assert loads('x:"a"  # comment') == Bind('a', name='x')
-    assert loads('x: "a"') == Bind('a', name='x')
-    assert loads('x : "a"') == Bind('a', name='x')
-
-
-def Grm(dfns):
-    return Grammar(definitions=dfns, start=next(iter(dfns)))
+    assert eloads('x:"a"') == Bind('a', name='x')
+    assert eloads('x:"a"  # comment') == Bind('a', name='x')
+    assert eloads('x: "a"') == Bind('a', name='x')
+    assert eloads('x : "a"') == Bind('a', name='x')
 
 
 def test_loads_def():
-    assert loads('A <- "a"') == Grm({'A': Literal('a')})
-    assert loads('A <- "a"  # comment') == Grm({'A': Literal('a')})
-    assert loads('A <- "a" "b"') == Grm({'A': Sequence('a', 'b')})
-    assert loads('A <- "a" B <- "b"') == Grm({'A': Literal('a'),
-                                              'B': Literal('b')})
+    assert loads('A <- "a"') == ('A', {'A': Literal('a')})
+    assert loads('A <- "a"  # comment') == ('A', {'A': Literal('a')})
+    assert loads('A <- "a" "b"') == ('A', {'A': Sequence('a', 'b')})
+    assert loads('A <- "a" B <- "b"') == ('A', {'A': Literal('a'),
+                                                'B': Literal('b')})
     assert loads('''
         A   <- "a" Bee
         Bee <- "b"
-    ''') == Grm({'A': Sequence('a', Nonterminal('Bee')),
-                 'Bee': Literal('b')})
+    ''') == ('A', {'A': Sequence('a', Nonterminal('Bee')),
+                   'Bee': Literal('b')})